Fixed Hampter being too small in portrait mode
FIX: Fixed the image used for Hacker Hampter being too small in portrait mode. Note that the image size does not update in real time, and only determines the size depending on the mode as soon as the button is pressed, and no other time unless the page is reloaded.